script: Move doc structures to an internal header
