+++ libs/script/src/doc.c
@@ -1,38 +1,8 @@
 #include "core_alloc.h"
 #include "core_diag.h"
 #include "core_dynarray.h"
-#include "script_doc.h"
 
-typedef u32 ScriptValId;
-
-typedef struct {
-  ScriptValId valId;
-} ScriptExprValue;
-
-typedef struct {
-  StringHash key;
-} ScriptExprLoad;
-
-typedef struct {
-  ScriptExpr  lhs;
-  ScriptExpr  rhs;
-  ScriptOpBin op;
-} ScriptExprOpBin;
-
-typedef struct {
-  ScriptExprType type;
-  union {
-    ScriptExprValue data_value;
-    ScriptExprLoad  data_load;
-    ScriptExprOpBin data_op_bin;
-  };
-} ScriptExprData;
-
-struct sScriptDoc {
-  DynArray   exprs;  // ScriptExprData[]
-  DynArray   values; // ScriptVal[]
-  Allocator* alloc;
-};
+#include "doc_internal.h"
 
 static ScriptExpr script_doc_expr_add(ScriptDoc* doc, const ScriptExprData data) {
   const ScriptExpr expr                         = (ScriptExpr)doc->exprs.size;
